requirements.txt
pandas
altair
vegafusion
vl-convert-python
shiny
shinywidgets
folium
geopandas
fiona
matplotlib
plotly
python-dotenv
querychat
chatlas
duckdb
pytest
pytest-playwright